requirements.txt
fastapi[standard]>=0.68.0
uvicorn>=0.15.0
moviepy==1.0.3
python-multipart>=0.0.6
python-dotenv>=0.19.0
replicate>=0.11.0
openai>=1.0.0
celery>=5.0.0
redis>=5.2.1
supabase>=1.0.0
python-jose[cryptography]>=3.3.0
requests>=2.25.1
openai-whisper
pexels-api